+++ src/pre.ts
@@ -1,0 +1,3 @@
+import treeGraphCommand from "./pre/tree-graph-command.js";
+
+export { treeGraphCommand };